Feat : 비밀번호 수정, 주소지관리 기능 추가

+++ src/main/java/com/sh/onezip/product/dto/ProductPurchaseInfoDto.java
@@ -13,7 +13,7 @@ public class ProductPurchaseInfoDto {
     private String productName;
     private int productQuantity;
     private int productPrice;
-    private int discountRate;
+    private double discountRate;
     private int totalProductPrice;
     private int totalDiscountPrice;
     private int totalOptionCost;
